+++ disconf-web/src/main/java/com/baidu/disconf/web/service/user/dao/UserDao.java
@@ -1,14 +1,14 @@
 package com.baidu.disconf.web.service.user.dao;
 
 import com.baidu.disconf.web.service.user.bo.User;
-import com.baidu.ub.common.generic.dao.BaseDao;
+import com.baidu.unbiz.common.genericdao.dao.BaseDao;
 
 /**
  * 
  * @author liaoqiqi
  * @version 2013-11-28
  */
-public interface UserDao extends BaseDao<Integer, User> {
+public interface UserDao extends BaseDao<Long, User> {
 
     void executeSql(String sql);
